\name{BASS-package}
\alias{BASS-package}
\alias{BASS}
\docType{package}
\title{
  Bayesian Analytics for Spatial Segmentation
}
\description{
  This package implements the BASS method for multi-scale and multi-sample 
  analysis in spatial transcriptomics.
}
\details{
  BASS performs multi-scale transcriptomic analyses in the form of joint cell 
  type clustering and spatial domain detection, with the two analytic tasks 
  carried out simultaneously within a Bayesian hierarchical modeling framework. 
  For both analyses, BASS properly accounts for the spatial correlation 
  structure and seamlessly integrates gene expression information with spatial 
  localization information to improve their performance. In addition, BASS is 
  capable of multi-sample analysis that jointly models multiple tissue 
  sections/samples, facilitating the integration of spatial transcriptomic 
  data across tissue samples.
}
\author{
Zheng Li, Xiang Zhou

Maintainer: Zheng Li <zlisph@umich.edu>
}
\references{
  Multi-scale and multi-sample analysis enables accurate cell type clustering
  and spatial domain detection in spatial transcriptomic studies
}
\keyword{ package }
\seealso{
  \href{https://zhengli09.github.io/BASS-Analysis/index.html}{BASS vignettes}
}
