\docType{package}
\name{bass}
\alias{bass}
\alias{bass-package}
\title{Tools and Analysis Functions for the Bering-Arctic Seal Surveys}
\description{
  bass is a central package for creating basic/global
  functions and data related to the Bering-Arctic Seal
  Surveys. The 2012 BOSS surveys are the first surveys to
  rely on this package
}
\author{
  Paul Conn & Josh London
}

