% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bassPCA.R
\name{bassPCA}
\alias{bassPCA}
\title{Bayesian Adaptive Spline Surfaces (BASS) with PCA decomposition}
\usage{
bassPCA(xx = NULL, y = NULL, dat = NULL, n.pc = NULL,
  perc.var = 99, n.cores = 1, center = T, scale = F, ...)
}
\arguments{
\item{xx}{a data frame or matrix of predictors.  Categorical predictors should be included as factors.}

\item{y}{a response matrix (functional response).}

\item{dat}{list with elements xx (same as above), y (same as above), n.pc (number of principal components used), basis (principal components), newy (reduced dimension y), trunc.error (truncation error), y.m (mean removed before PCA), y.s (sd scaled before PCA)}

\item{...}{areguements to be passed to bass.}
}
\value{
An object of class 'bass'.  The other output will only be useful to the advanced user.  Rather, users may be interested in prediction and sensitivity analysis, which are obtained by passing the entire object to the predict.bass or sobol functions.
}
\description{
Fits a BASS models in PCA space using RJMCMC.
}
\details{
Fits a bass model to each basis coefficient, \code{bass(dat$xx,dat$newy[i,],...)} for \code{i in 1 to n.pc}.
}
\seealso{
\link{predict.bass} for prediction and \link{sobol} for sensitivity analysis.
}
\keyword{analysis}
\keyword{data}
\keyword{functional}
\keyword{nonparametric}
\keyword{regression,}
\keyword{splines,}
