% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS.R
\name{evalLik}
\alias{evalLik}
\title{Evaluate posterior density at each interation assuming
beta is fixed}
\usage{
evalLik(BASS)
}
\description{
Evaluate posterior density at each interation assuming
beta is fixed
}
