\name{get.edge.Bering}
\alias{get.edge.Bering}
\title{calculate southern ice edge from spatial polygons object giving sea ice concentration}
\usage{
  get.edge.Bering(SpDF, proj)
}
\arguments{
  \item{SpDF}{SpatialPolygonsDataFrame object giving sea
  ice concentration}

  \item{proj}{projection associated with the object}
}
\value{
  SpatialLines object giving approx southern ice edge
}
\description{
  calculate southern ice edge from spatial polygons object
  giving sea ice concentration
}
\author{
  Paul Conn \email{paul.conn@noaa.gov}
}

