% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS.R
\name{listAllHyper}
\alias{listAllHyper}
\title{Show BASS settings}
\usage{
listAllHyper(BASS)
}
\arguments{
\item{BASS}{A BASS object created by \code{createBASSObject}.}
}
\description{
Show a complete list of settings, hyper-parameters, and MCMC controlling
parameters used by BASS.
}
