% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc_coinflip.R
\name{makeCoinWeights}
\alias{makeCoinWeights}
\title{Make Weights for N2KD (coinflip) Proposal}
\usage{
makeCoinWeights(eta, p0, w3)
}
\arguments{
\item{eta}{vector of counts. The ith element of eta denotes the number of times predictor i is currently included in the model}

\item{p0}{The expected interaction order}

\item{alpha}{Tuning parameter (non-negative). Larger values indicate more preference for previously selected variables (0 is uniform).}

\item{epsilon}{Tuning parameter (non-negative). A laplace smoother, larger values lead to more uniform sampling.}

\item{num_passes}{Number of passes for the deterministic weight generator.}
}
\value{
A vector of coin flip weights
}
\description{
Finds probabilities for each coins based on count vector and tuning parameters
}
\details{
Generates coin flip weights
}
\examples{
p <- 20
p0 <- 8
eta <- rep(1, p)
w <- make_weights(eta, p0)
sum(w)
w

eta[1:3] <- 1e5
w <- make_weights(eta, p0)
sum(w)
w

eta <- abs(1e3*rt(p, 3))
w <- make_weights(eta, p0)
sum(w)
w
}
