% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bassBasis}
\alias{plot.bassBasis}
\title{BASS Plot Diagnostics}
\usage{
\method{plot}{bassBasis}(x, quants = c(0.025, 0.975), pred = T, ...)
}
\arguments{
\item{x}{a \code{bassBasis} object.}

\item{quants}{quantiles for intervals, if desired.  NULL if not desired.}

\item{pred}{logical, should predictive performance be plotted?}

\item{...}{graphical parameters.}
}
\description{
Generate diagnostic plots for BASS model fit.
}
\details{
The first two plots are trace plots for diagnosing convergence.  The third plot is posterior predicted vs observed, with intervals for predictions.  The fourth plot is a histogram of the residuals (of the posterior mean model).  If \code{pred = FALSE}, the third and fourth plots are omitted.
}
\examples{
# See examples in bassBasis documentation.

}
\seealso{
\link{bassBasis}, \link{bassPCA}, \link{predict.bassBasis}, \link{sobolBasis}
}
