% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plotClusters}
\alias{plotClusters}
\title{plot cell types or tissue structures}
\usage{
plotClusters(
  xy,
  labels,
  title = "",
  point_size = 1,
  flip_x = F,
  flip_y = F,
  flip_xy = F,
  ratio = 1
)
}
\description{
plot cell types or tissue structures
}
