% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plotGene}
\alias{plotGene}
\title{plot gene expression values}
\usage{
plotGene(
  xys,
  cnts,
  log = T,
  scale = F,
  title = "",
  legend = "",
  point_size = 1,
  flip_x = FALSE,
  flip_y = FALSE,
  flip_xy = FALSE
)
}
\description{
plot gene expression values
}
