% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plotMkrs}
\alias{plotMkrs}
\title{expression of marker genes stratified by cell types}
\usage{
plotMkrs(cnts, markers, labels, legend_title, scale = T)
}
\description{
expression of marker genes stratified by cell types
}
