% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.bass}
\alias{print.bass}
\title{Print BASS Details}
\usage{
\method{print}{bass}(x, ...)
}
\arguments{
\item{x}{a \code{bass} object, returned from \code{bass}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print some of the details of a BASS model.
}
