% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.bassBasis}
\alias{print.bassBasis}
\title{Print BASS Details}
\usage{
\method{print}{bassBasis}(x, ...)
}
\arguments{
\item{x}{a \code{bassBasis} object, returned from \code{bassPCA} or \code{bassBasis}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print some of the details of a BASS model.
}
