% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_taylor.R
\name{shapley_taylor}
\alias{shapley_taylor}
\title{Shapley-Taylor Effects with BASS}
\usage{
shapley_taylor(
  bassSob,
  max_order = 3,
  relative_importance = function(j) rep(1, length(j)),
  proportion = FALSE,
  mcmc.use = NULL,
  verbose = TRUE
)
}
\arguments{
\item{bassSob}{output from the \code{sobol} function.}

\item{max_order}{Generalize shapley effects will be computed for all interaction orders up to \code{maxOrder}.}

\item{relative_importance}{a function with one argument (j) giving the (unnormalized) relative importance for subsets of size j. Default is function(j) 1.}

\item{proportion}{should shapley effects be returned on variance or proportion scale?}

\item{mcmc.use}{an integer vector indexing which MCMC iterations to use for sensitivity analysis.}

\item{verbose}{logical; should progress be displayed?}
}
\value{
A matrix containing estimated posteriors for the Shapley values.
}
\description{
Assess the sensitivity of a computer model using Shapley values
}
\details{
Calculates the Shapley values based on the Sobol decomposition
}
\examples{
f<-function(x){
10*sin(2*pi*x[,1]*x[,2])+20*(x[,3]-.5)^2+10*x[,4]+5*x[,5]
}
sigma<-1 # noise sd
n<-50 # number of observations
x<-matrix(runif(n*10),n,10) #10 variables, only first 5 matter
y<-rnorm(n,f(x),sigma)

mod <- bass(x, y)
sob <- sobol(mod)
shapley(mod)

}
\seealso{
\link{bass} for model fitting and \link{sobol} for Sobol decomposition.
}
\keyword{Shapley}
\keyword{value}
