% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bassPCA.R
\name{sobolOB}
\alias{sobolOB}
\title{BASS Sensitivity Analysis}
\usage{
sobolOB(mod, int.order, prior = NULL, mcmc.use = 1, nind = NULL,
  ncores = 1, preschedule = F, plot = F)
}
\arguments{
\item{mod}{output from the \code{bassOB} or \code{bassPCA} function.}

\item{int.order}{an integer indicating the highest order of interactions to include in the Sobol decomposition.}

\item{prior}{a list with the same number of elements as there are inputs to mod.  Each element specifies the prior for the particular input.  If unspecified, a uniform is assumed with the same bounds as are represented in the input to xx.}

\item{mcmc.use}{an integer vector indexing which MCMC iterations to use for sensitivity analysis.}

\item{nind}{number of Sobol indices to keep (will keep the largest nind).}

\item{ncores}{number of cores to use (nearly linear speedup for adding cores).}

\item{preschedule}{to be passed to mclapply.}

\item{plot}{logical; whether to plot results.}
}
\value{
If non-functional (\code{func.var = NULL}), a list with two elements:
 \item{S}{a data frame of sensitivity indices with number of rows matching the length of \code{mcmc.use}.  The columns are named with a particular main effect or interaction.  The values are the proportion of variance in the model that is due to each main effect or interaction.}
 \item{T}{a data frame of total sensitivity indices with number of rows matching the length of \code{mcmc.use}.  The columns are named with a particular variable.}
 Otherwise, a list with four elements:
 \item{S}{an array with first dimension corresponding to MCMC samples (same length as \code{mcmc.use}), second dimension corresponding to different main effects and interactions (labeled in \code{names.ind}), and third dimension corresponding to the grid used for the functional variable.  The elements of the array are sensitivity indices.}
 \item{S.var}{same as \code{S}, but scaled in terms of total variance rather than percent of variance.}
 \item{names.ind}{a vector of names of the main effects and interactions used.}
}
\description{
Decomposes the variance of the BASS model into variance due to main effects, two way interactions, and so on, similar to the ANOVA decomposition for linear models.  Uses the Sobol' decomposition, which can be done analytically for MARS models.
}
\details{
Performs analytical Sobol' decomposition for each MCMC iteration in mcmc.use (each corresponds to a MARS model), yeilding a posterior distribution of sensitivity indices.  Can obtain Sobol' indices as a function of one functional variable.
}
\examples{
# See examples in bass documentation.
}
\seealso{
\link{bass} for model fitting and \link{predict.bass} for prediction.
}
\keyword{Sobol}
\keyword{decomposition}
