% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.bass}
\alias{summary.bass}
\title{Summarize BASS Details}
\usage{
\method{summary}{bass}(object, ...)
}
\arguments{
\item{object}{a \code{bass} object, returned from \code{bass}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarize some of the details of a BASS model.
}
