% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.bassBasis}
\alias{summary.bassBasis}
\title{Summarize BASS Details}
\usage{
\method{summary}{bassBasis}(object, ...)
}
\arguments{
\item{object}{a \code{bassBasis} object, returned from \code{bassPCA} or \code{bassBasis}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarize some of the details of a BASS model.
}
