% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{writeBassJSON}
\alias{writeBassJSON}
\alias{readBassJSON}
\title{Write/read BASS object to JSON}
\usage{
writeBassJSON(x, file, ...)

readBassJSON(file)
}
\arguments{
\item{x}{an object of class \code{bass}, \code{bassBasis}, or \code{bassSob}, returned from \code{bass}, \code{bassPCA}, \code{bassBasis}, \code{sobol}, or \code{sobolBasis}.}

\item{file}{json file to write or read.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Write/read an object of class \code{bass}, \code{bassBasis}, or \code{bassSob} to a JSON file.
}
