% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS.R
\docType{class}
\name{BASS-class}
\alias{BASS-class}
\title{BASS object}
\description{
BASS object stores data, settings, hyper-parameters, MCMC controlling 
parameters, and results for the spatial transcriptomic analysis.
}
\section{Slots}{

\describe{
\item{\code{X}}{An L-list of raw gene expression count matrices for L tissue 
sections. For the lth tissue section, the matrix has dimension PxNl 
representing P common genes shared across tissue sections and Nl 
cells/spots in the lth tissue section.}

\item{\code{X_run}}{A JxN gene expression feature matrix obtained from 
\code{BASS.preprocess} for running the BASS algorithm. Columns are cells 
from all tissue sections and rows are low-dimensional gene expression 
features. Modify this slot if you have pre-processed the data by yourself.}

\item{\code{xy}}{An L-list of spatial coordinates matrices for L tissue sections.
For the lth tissue section, the matrix has dimension Nlx2 representing the 
same Nl cells/spots as in the expression count matrix and their x and y 
spatial coordinates.}

\item{\code{L}}{Numeric. Number of tissue sections.}

\item{\code{P}}{Numeric. Number of genes.}

\item{\code{Ns}}{Numeric vector. Number of cells/spots in each of the L tissue 
sections.}

\item{\code{C}}{Numeric. Number of cell types.}

\item{\code{R}}{Numeric. Number of spatial domains.}

\item{\code{init_method}}{Character. Initialize the cell type clusters and spatial 
domains with either k-means clustering (\code{kmeans}) or Gaussian mixture 
model (\code{mclust}).}

\item{\code{psi0}}{Numeric. psi0*I is the scale matrix of the inv-Wishart prior on 
the variance-covariance matrix, where I is an identity matrix.}

\item{\code{n0}}{Numeric. Degrees of freedom of the inv-Wishart prior on the 
variance-covariance matrix.}

\item{\code{alpha0}}{Numeric. Concentration parameter of the Dirichlet prior
specified on the cell type composition vector in each spatial domain.}

\item{\code{k}}{Numeric. Minimum number of neighbors for each cell/spot based on 
the Euclidean distance.}

\item{\code{burnin}}{Numeric. Number of burn-in iterations in the MCMC.}

\item{\code{nsample}}{Numeric. Number of posterior samples in the MCMC.}

\item{\code{beta_method}}{Character. Fix the cell-cell interaction parameter 
to be \code{beta} (\code{fix}) or estimate the parameter based on the 
data at hand (\code{SW}).}

\item{\code{beta}}{Numeric. Pre-specified cell-cell interaction parameter if 
\code{beta_method} is set to be \code{fix} or otherwise its initial value.}

\item{\code{step_size}}{Numeric. Step size of a uniform random walk.}

\item{\code{tol}}{Numeric. Threshold of convergence for the cell-cell 
interaction parameter.}

\item{\code{potts_burnin}}{Numeric. Number of burn-in iterations in Potts sampling.}

\item{\code{potts_nsample}}{Numeric. Number of Potts samples to approximate the 
partition ratio.}

\item{\code{samples}}{List. Samples from the MCMC. The list contains:
 \itemize{
   \item c: NxS matrix of cell type labels with cells/spots as rows and 
     posterior samples as columns. Cells/spots across all tissue sections 
     have been stacked into a vector.
   \item z: NxS matrix of spatial domain labels with cells/spots as rows 
     and posterior samples as columns. Cells/spots across all tissue 
     sections have been stacked into a vector.
   \item beta: Numeric. Burn-in samples of the cell-cell interaction 
     parameter \code{beta}.
   \item pi: CxRxS array with each element (c, r, s) representing the 
     proportion of cell type c in spatial domain r in the posterior 
     sample s.
   \item mu: JxCxS array with each element (j, c, s) representing the mean 
     of the jth feature in cell type c in the posterior sample s. 
   \item Sigma: JxJxS array with each slice s representing the JxJ 
     variance-covariance matrix in the gene expression model in the 
     posterior sample s.
   \item d: JxS matrix with each element (j, s) representing the jth 
     feature-specific prior mean of mu_jc in the posterior sample s.
   \item lambda: JxS matrix with each element (j, s) representing the jth 
     feature-specific scaling factor of mu_jc in the posterior sample s.
   \item loglik: log-likelihood in each MCMC iteration.
 }
Refer to the supplementary file of the paper for details of all the 
parameters.}

\item{\code{results}}{List. Estimates of cell type labels, spatial domain 
labels, and cell type composition in each spatial domain after 
post-processing the posterior samples in \code{samples}.
The list contains:
\itemize{
  \item init_c: N-vector of initial cell type labels obtained using the 
    method specified by \code{init_method}. Cells/spots across all tissue 
    sections have been stacked into a vector.
  \item init_z: N-vector of initial spatial domain labels obtained using 
    the method specified by \code{init_method}. Cells/spots across all 
    tissue sections have been stacked into a vector.
   \item c: List. Cell type labels for each of the L tissue sections.
   \item z: List. Spatial domain labels for each of the L tissue sections.
   \item p: Numeric matrix. CxR matrix with the cth row and rth column 
     representing the proportion of cell type c in the spatial domain r.
  \item beta: Numeric. Estimate of the cell-cell interaction parameter 
    \code{beta}.
  \item loglik: log-likelihood.
}}

\item{\code{elapsed_time:}}{Elapsed time for running MCMC in seconds.}
}}

