% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS.R
\docType{class}
\name{BASS-class}
\alias{BASS-class}
\title{BASS object stores all the data and parameters used to run the algorithm}
\description{
BASS object stores all the data and parameters used to run the algorithm
}
\section{Slots}{

\describe{
\item{\code{X}}{L-list of raw gene expression matrix for L samples}

\item{\code{X_run}}{N x J matrix of processed expression matrix for running BASS}

\item{\code{xy}}{L-list of location matrix for L samples}

\item{\code{L}}{Number of samples}

\item{\code{P}}{Number of genes}

\item{\code{Ns}}{L-list of number of cells/spots for L samples}

\item{\code{C}}{Number of cell types}

\item{\code{R}}{Number of spatial domains}

\item{\code{init_method}}{Initialization method for both cell types and 
tissue structures}

\item{\code{cov_struc}}{Covariance structure for gene expression features}

\item{\code{kappa}}{Prior parameter of gene expression}

\item{\code{alpha0}}{Concentration parameter of Dirichlet distribution}

\item{\code{a}}{Shape parameter of the inverse gamma prior}

\item{\code{b}}{Scale parameter of the inverse gamma prior}

\item{\code{W0}}{W0*I represents the scale matrix of Wishart prior}

\item{\code{n0}}{Degrees of freedom of Wishart prior}

\item{\code{k}}{Number of neighbors for kNN graph}

\item{\code{burn_in}}{Number of burn-in iterations}

\item{\code{samples}}{Number of posterior samples}

\item{\code{beta_est_approach}}{Approach to estimate beta ("FIXED", "ACCUR_EST")}

\item{\code{beta}}{Fixed value of beta if the betaEstApproach is set to be "FIXED"}

\item{\code{beta_max}}{upper bound of beta}

\item{\code{epsilon}}{Uniform random walk step size}

\item{\code{beta_tol}}{tolerance for checking beta convergence}

\item{\code{B}}{Number of burn-in steps in Potts sampling}

\item{\code{M}}{Number of Potts samples to approximate Potts partition ratio}

\item{\code{res}}{Result list}

\item{\code{res_postprocess}}{Result list after post-processing}
}}

