% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS.R
\name{BASS.preprocess}
\alias{BASS.preprocess}
\title{Preprocess expression data}
\usage{
BASS.preprocess(
  BASS,
  doLogNormalize = TRUE,
  geneSelect = c("sparkx", "hvgs"),
  doPCA = TRUE,
  scaleFeature = TRUE,
  doBatchCorrect = TRUE,
  nHVG = 2000,
  nSE = 3000,
  nPC = 20
)
}
\description{
Preprocess expression data
}
