% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS.R
\name{createBASSObject}
\alias{createBASSObject}
\title{Create BASS object}
\usage{
createBASSObject(
  X,
  xy,
  C,
  R,
  init_method = c("kmeans", "mclust"),
  cov_struc = c("EEE", "EII"),
  kappa = 0.01,
  alpha0 = 1,
  a = 2,
  b = 1,
  W0 = 1,
  n0 = 1,
  k = 4,
  burn_in = 10000,
  samples = 10000,
  beta_est_approach = c("FIXED", "ACCUR_EST"),
  beta = 1,
  beta_max = 4,
  epsilon = 0.1,
  beta_tol = 0.1,
  B = 100,
  M = 100
)
}
\description{
Create BASS object
}
